WIP- #57 Refactor audio backends to separate module
